chore: move tweets feed to Tweets.js
